Refactor TOC handling and support nested TOC configurations
Introduced `TableOfContentsConfiguration` to refactor how TOC files are parsed and improve code modularity. Added support for a `max_toc_depth` parameter to restrict TOC nesting depth. Updated configurations and comments to clarify support for nested TOCs and modified file paths accordingly.
