+++ docs/development/link-validation/toc.yml
@@ -1,0 +1,2 @@
+toc:
+  - file: link-validation.md
