+++ docs/development/toc.yml
@@ -1,3 +1,3 @@
 toc:
   - file: index.md
-  - file: link-validation.md
+  - toc: link-validation
